VBO - Cone, cilindro
->Só falta vbo do plano

+++ Preview do GERADOR/CG-Aula01/anel.h
@@ -24,7 +24,7 @@ public:
     
 protected:
     GLuint buffers[1];
-    int n_indices;
+    short n_indices;
     unsigned short *indices;
 };
